requirements.txt
# Hearth Writer v1.2 "Vertex" - Hermetically Sealed Dependencies
# Generated: 2025-01-03
# Install: pip install -r requirements.txt
# 
# IMPORTANT: All versions are pinned for reproducibility.
# Do not use >= or ~= operators in production.

# === Core Framework ===
flask==3.0.0
flask-sock==0.7.0
werkzeug==3.0.1
psutil==5.9.7

# === AI/ML Inference (Local) ===
llama-cpp-python==0.2.56
openai-whisper==20231117
TTS==0.22.0

# === Vector Database (RAG) ===
lancedb==0.4.1
pyarrow==14.0.2

# === Data Handling ===
numpy==1.26.3
scipy==1.11.4
soundfile==0.12.1

# === Text Processing ===
regex==2023.12.25
tiktoken==0.5.2

# === Sync Protocol ===
watchdog==3.0.0

# === Utilities ===
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1

# === Development/Testing ===
pytest==7.4.4
pytest-asyncio==0.23.3

# === Transitive Dependencies (Pinned for Hermetic Seal) ===
click==8.1.7
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
simple-websocket==1.0.0
wsproto==1.2.0

# === Optional: GPU Acceleration ===
# Uncomment ONE of the following based on your hardware:
#
# For NVIDIA CUDA (requires CUDA toolkit 11.8+):
# torch==2.1.2+cu118
# --extra-index-url https://download.pytorch.org/whl/cu118
#
# For Apple Silicon (Metal):
# CMAKE_ARGS="-DLLAMA_METAL=on" pip install llama-cpp-python==0.2.56
#
# For CPU-only (default, no changes needed)
